pyproject.toml
[project]
name = "souk-dz-agent"
version = "0.1.0"
description = "Daily price-arbitrage agent monitoring Algerian marketplaces, Facebook, and TikTok"
authors = [{ name = "Hadil Aboubaker" }]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    # HTTP + parsing
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "tenacity>=9.0.0",
    # Browser automation (for SPAs: Ouedkniss, Facebook, TikTok)
    "playwright>=1.49.0",
    # AI
    "google-generativeai>=0.8.0",
    # Reporting
    "openpyxl>=3.1.5",
    "jinja2>=3.1.4",
    # Config + utils
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.0",
    "typer>=0.15.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[project.scripts]
souk-dz = "souk_dz.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["souk_dz"]

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]
